pyproject.toml
[tool.poetry]
name = "otel-python-grpc-nested-spans"
version = "0.1.0"
description = ""
authors = ["Eric Hulburd <ehulburd@rigetti.com>"]
readme = "README.md"
packages = [{include = "otel_python_grpc_nested_spans"}]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.50.0"
grpcio-tools = "^1.50.0"
opentelemetry-api = "^1.14.0"
opentelemetry-sdk = "^1.14.0"
opentelemetry-instrumentation-grpc = "^0.35b0"
opentelemetry-exporter-jaeger-proto-grpc = "^1.14.0"
numpy = "^1.23.4"
grpcio-health-checking = "^1.50.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
